Move usershortcuts to canteen package
